refactor: switched on strictest typing ruleset and fixed resulting errors

+++ spin-onehalf-square-hcbosons/hamiltonian.py
@@ -36,7 +36,7 @@ class Hamiltonian(ABC):
         self,
         time: float,
         system_state: state.SystemState,
-    ) -> float:
+    ) -> np.complex128:
         pass
 
     @abstractmethod
